Remove Loop Guards
Previously -- when addCallback was on Object.prototype -- all objects had a member that was a function.  Therefore, in order to make traversing the properties of an object work as intended, it was necessary for such loops to have a guard that skipped properties that were functions.

Since addCallback is now CivicInfo.AddCallback, this is no longer necessary.