removeClass Fix
Fixed a bug that would occasionally cause removeClass to corrupt the CSS class of the element-in-question.

This occurred because the regex it used to delete the appropriate class matched whitespace or end/beginning of string on both sides, meaning that if a class in the middle (say "foo" from "bar foo baz") was removed, the classes surrounding the removed class would become merged (resulting in "barbaz"), changing their meaning.

Corrected this by changing one of the matches to a lookahead.