ForEach
Wrote a ForEach function, which traverses some collection, invoking a callback for each item.

If any callback returns anything, that value is returned and the looping stops immediately, otherwise the looping goes to completion and nothing is returned.

If the function is passed something that isn't a collection (i.e. doesn't have a length property) then the callback is invoked just once for that single item, and the return value (if any) thereof is returned.