Form Validation
Added CivicInfoBC.Form, which provides convenient, markup-based form validation.  One simply encapsulates a form within a CivicInfoBC.Form, and then uses HTML5-compliant markup to describe the conditions under which the form is considered to be invalid.

data-mandatory specifies that a field must be provided.

data-mandatory-if specifies that a field must be provided contingent upon some JavaScript function call (the function must be a member of the object doing the validation).

data-regex specifies that a field is valid if it matches some regular expression.

data-callback specifies that a field is valid contigent on the result of some JavaScript function call (the function must be a member of the object doing the validation).

Radio buttons are mandatory by default.

Checkboxes are not checked at present.